Authentication with APIs : JWT

+++ controllers/api/v1/posts_api.js
@@ -22,18 +22,21 @@ module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
 
-    // if (post.user == req.user.id) {
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({ post: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    return res.json(200, {
-      message: 'Post and associated comments deleted!',
-    });
-    // } else {
-    //   return res.redirect('back');
-    // }
+      return res.json(200, {
+        message: 'Post and associated comments deleted!',
+      });
+    } else {
+      return res.json(401, {
+        message: 'You cannot delete this post',
+      });
+    }
   } catch (err) {
+    console.log(err);
     req.flash('error', err);
     return res.json(500, {
       message: 'Internal server error',
